+++ precompiles/ics20/ics20.go
@@ -1,7 +1,7 @@
 package ics20
 
 import (
-	"embed"
+	"bytes"
 	"fmt"
 
 	"github.com/ethereum/go-ethereum/accounts/abi"
@@ -14,10 +14,9 @@ import (
 	storetypes "cosmossdk.io/store/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-)
 
-// PrecompileAddress of the ICS-20 EVM extension in hex format.
-const PrecompileAddress = "0x0000000000000000000000000000000000000802"
+	_ "embed"
+)
 
 var _ vm.PrecompiledContract = &Precompile{}
 
@@ -25,13 +24,13 @@ var (
 	// Embed abi json file to the executable binary. Needed when importing as dependency.
 	//
 	//go:embed abi.json
-	f   embed.FS
+	f   []byte
 	ABI abi.ABI
 )
 
 func init() {
 	var err error
-	ABI, err = cmn.LoadABI(f, "abi.json")
+	ABI, err = abi.JSON(bytes.NewReader(f))
 	if err != nil {
 		panic(err)
 	}