+++ precompiles/werc20/werc20.go
@@ -1,7 +1,7 @@
 package werc20
 
 import (
-	"embed"
+	"bytes"
 	"slices"
 
 	"github.com/ethereum/go-ethereum/accounts/abi"
@@ -13,6 +13,8 @@ import (
 	erc20types "github.com/cosmos/evm/x/erc20/types"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	_ "embed"
 )
 
 // abiPath defines the path to the WERC-20 precompile ABI JSON file.
@@ -22,13 +24,13 @@ var (
 	// Embed abi json file to the executable binary. Needed when importing as dependency.
 	//
 	//go:embed abi.json
-	f   embed.FS
+	f   []byte
 	ABI abi.ABI
 )
 
 func init() {
 	var err error
-	ABI, err = cmn.LoadABI(f, abiPath)
+	ABI, err = abi.JSON(bytes.NewReader(f))
 	if err != nil {
 		panic(err)
 	}