+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### IMPROVEMENTS
 
+- [\#]() Cleanup precompiles abi.json.
+
 ### FEATURES
 
 ### BUG FIXES